backend/requirements.txt
fastapi>=0.120.0
uvicorn>=0.38.0
python-dotenv>=1.0.0
pydantic>=2.12.0
sqlalchemy>=1.4.23
python-multipart>=0.0.20
requests>=2.32.5
Pillow>=10.3.0